+++ packages/utils/src/index.ts
@@ -1,3 +1,4 @@
+export { arrayContentEquals } from "./arrays";
 export { assert, assertDefined } from "./assert";
 export { sleep } from "./sleep";
 export { isNonNullObject, isUint8Array } from "./typechecks";